move skycell wcs reading into resample
